Update Main / packages / README
Update README
Update Main
Nova pasta packages:
- eval_module
- login_module
- signup_module
Add product backlog
Add burndown
Main funcional (básico)
Separei arquivos originais em eval_test para criação de futuras funções
