--- Cadastro/cadastro.csv
@@ -1,11 +1,0 @@
-id,nome,email,senha
-1001,joao,aaa,111
-1002,teresa,bbb,222
-1003,maria,ccc,333
-1004,carlos,ddd,444
-1005,larissa,eee,555
-1006,geoge,fff,666
-1007,kaio,ggg,777
-1008,beatriz,hhh,888
-1009,igor,iii,999
-1010,yasmin,jjj,000
