+++ SprintLogs/Fluxogramas/Fluxogramas.md
@@ -1,0 +1,7 @@
+### Fluxogramas
+
+###### (Está sujeito à modificações)
+- Diagrama de Acessos
+<img src="Login.png" alt="Login" width="750"/>
+- Fluxograma do Banco de Dados
+<img src="DiagramaDB1.0.png" alt="DiagramaDB1.0" width="750"/>
