--- SprintLogs/SPRINT1.md
@@ -1,31 +1,0 @@
-## Observações da Sprint 1
-
-Devido ao aprendizado com as metodologias ageis, o foco em grande parte foi em relação ao planejamento e organização do grupo e do projeto. A documentação base é o essencial, assim como a definição das prioridades para o desenvolvimento, cliente, e a busca por clareza dos objetivos.
-
-#### Registros do planejamento:
-- [Brainstorming Board (Miro)](https://miro.com/app/board/uXjVMe5mTZ4=/?share_link_id=136549221515)
-- [Board Scrum (Miro)](https://miro.com/app/board/uXjVMY5EpQI=/?share_link_id=258031151867)
-- [Distribuição de Tasks (ClickUp)](https://sharing.clickup.com/9007095918/l/h/4-90070215479-1/fbcc296d506a261)
-
-### Fluxogramas
-
-- Diagrama de Acessos
-
-<img src="Login.png" alt="Login" width="750"/>
-
-- Fluxograma do Banco de Dados
-
-<img src="DiagramaDB1.0.png" alt="DiagramaDB1.0" width="750"/>
-
-# Entregáveis da Sprint 1
-> Expectativas para sprint após o planejamento, não seguem uma ordem em particular nesta representação.
-
-| Descrição | Prioridade |
-| :---- | :----: |
-| Documentação da Sprint Planning | Essencial |
-| Registro das tecnologias/ferramentas utilizadas | Essencial |
-| Integração dos membros da equipe no projeto | Essencial |
-| Inclusão do backlog do produto | Essencial |
-| Documentação backlog das Sprints | Essencial |
-| Apresentação do wireframe | Essencial |
-| Definição de MVP | Importante |
