+++ data_sample/aval_data.csv
@@ -1,0 +1,6 @@
+id,nome,turma,time,sprint,criterio,s1c1,s1c2,s1c3,s1c4,s1c5,s2c1,s2c2,s2c3,s2c4,s2c5,s3c1,s3c2,s3c3,s3c4,s3c5,s4c1,s4c2,s4c3,s4c4,s4c5
+1001,joao,1,1,1,1,,,,,,,,,,,,,,,,,,,,,
+1002,maria,2,1,1,2,,,,,,,,,,,,,,,,,,,,,
+1003,daniel,1,2,1,3,,,,,,,,,,,,,,,,,,,,,
+1004,yasmin,1,1,1,4,,,,,,,,,,,,,,,,,,,,,
+1005,julia,1,1,1,5,,,,,,,,,,,,,,,,,,,,,
