+++ requirements.txt
@@ -1,0 +1,31 @@
+asttokens==2.2.1
+backcall==0.2.0
+colorama==0.4.6
+comm==0.1.3
+debugpy==1.6.7
+decorator==5.1.1
+executing==1.2.0
+ipykernel==6.22.0
+ipython==8.12.0
+jedi==0.18.2
+jupyter_client==8.2.0
+jupyter_core==5.3.0
+matplotlib-inline==0.1.6
+nest-asyncio==1.5.6
+packaging==23.1
+parso==0.8.3
+pickleshare==0.7.5
+platformdirs==3.2.0
+prompt-toolkit==3.0.38
+psutil==5.9.5
+pure-eval==0.2.2
+pwinput==1.0.3
+Pygments==2.15.1
+python-dateutil==2.8.2
+pywin32==306
+pyzmq==25.0.2
+six==1.16.0
+stack-data==0.6.2
+tornado==6.3.1
+traitlets==5.9.0
+wcwidth==0.2.6
