--- sample.csv
@@ -1,10 +1,0 @@
-nome,acesso,time,turma
-caio,integrante,N,1
-larissa,integrante,V,2
-george,integrante,N,2
-yara,integrante,V,2
-rafaela,integrante,N,1
-fabio,integrante,N,2
-wallace,adm,null,null
-isabela,integrante,N,1
-priscila,integrante,V,2
